feedback-agent/pyproject.toml
[tool.poetry]
name = "feedback-agent"
version = "0.1.0"
description = "FastAPI backend for the feedback AI agent"
authors = ["Nurlan <asankozheovn@gmail.com>", "Aidin <aidinror@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.3.13"
fastapi = "*"
uvicorn = "*"
langgraph = "*"
langchain = "*"
python-dotenv = "*"
requests = "*"
pydantic = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"